requirements.txt
flask
gunicorn
google-generativeai
pysqlite3
markdown
wikipedia